chore: Move bepu asset compilation to Stride.Assets
