+++ sources/editor/Stride.Assets.Presentation/Stride.Assets.Presentation.sdpkg
@@ -16,6 +16,7 @@ TemplateFolders:
       Files:
         - !file Templates/Assets/Animations/DefaultAnimation.sdtpl
         - !file Templates/Assets/Animations/DefaultSkeleton.sdtpl
+        - !file Templates/Assets/BepuPhysics/HullAsset.sdtpl
         - !file Templates/Assets/Fonts/OfflineRasterizedSpriteFont.sdtpl
         - !file Templates/Assets/Fonts/RuntimeRasterizedSpriteFont.sdtpl
         - !file Templates/Assets/Fonts/SignedDistanceFieldSpriteFont.sdtpl
