+++ sources/editor/Stride.Assets.Presentation/Templates/Assets/BepuPhysics/HullAsset.sdtpl
