+++ sources/editor/Stride.Editor/Stride.Editor.csproj
@@ -33,6 +33,7 @@
     <ProjectReference Include="..\..\engine\Stride.Navigation\Stride.Navigation.csproj" />
     <ProjectReference Include="..\..\engine\Stride.Particles\Stride.Particles.csproj" />
     <ProjectReference Include="..\..\engine\Stride.Physics\Stride.Physics.csproj" />
+    <ProjectReference Include="..\..\engine\Stride.BepuPhysics\Stride.BepuPhysics\Stride.BepuPhysics.csproj" />
     <ProjectReference Include="..\..\engine\Stride.SpriteStudio.Offline\Stride.SpriteStudio.Offline.csproj" />
     <ProjectReference Include="..\..\engine\Stride.UI\Stride.UI.csproj" />
     <ProjectReference Include="..\..\engine\Stride.Video\Stride.Video.csproj" />
