+++ sources/engine/Stride.BepuPhysics/Stride.BepuPhysics/Stride.BepuPhysics.csproj
@@ -11,7 +11,6 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <StrideAssemblyProcessor>true</StrideAssemblyProcessor>
     <StrideAssemblyProcessorOptions>--serialization --parameter-key</StrideAssemblyProcessorOptions>
-    <StridePackAssets>true</StridePackAssets>
     <StrideBuildTags>*</StrideBuildTags>
     <StrideCodeAnalysis>true</StrideCodeAnalysis>
   </PropertyGroup>
@@ -21,17 +20,9 @@
     <Compile Include="..\..\..\shared\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
-    <Reference Include="VHACDSharp">
-      <HintPath>..\..\..\..\deps\VHACD\VHACDSharp.dll</HintPath>
-    </Reference>
-    <StrideNativeLib Include="..\..\..\..\deps\VHACD\**\VHACD.*">
-      <Link>runtimes\%(RecursiveDir)native\%(Filename)%(Extension)</Link>
-    </StrideNativeLib>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\assets\Stride.Core.Assets\Stride.Core.Assets.csproj" />
-    <ProjectReference Include="..\..\Stride.Assets\Stride.Assets.csproj" />
     <ProjectReference Include="..\..\Stride.Engine\Stride.Engine.csproj" />
   </ItemGroup>
