--- sources/engine/Stride.BepuPhysics/Stride.BepuPhysics/Stride.BepuPhysics.sdpkg
@@ -1,7 +1,0 @@
-!Package
-SerializedVersion: {Assets: 3.1.0.0}
-TemplateFolders:
-    - Path: !dir Templates/Assets
-      Group: Assets
-      Files:
-        - !file Templates/Assets/Bepu/HullAsset.sdtpl
