--- sources/engine/Stride.BepuPhysics/Stride.BepuPhysics/Templates/Assets/.sdtpl/ColliderHull.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:81ea3240a736df6fcf27d8153af2f1fc1d360aa3bdffda65ccd45b5d5f884ed6
-size 2063
